# backend/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost
  description: Generated server url
paths:
...
  /api/users/me/avatar:
    post:
      tags:
      - user-controller
      operationId: uploadAvatar
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
              - file
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserSelfResponse"
      security:
      - bearerAuth: []
...
  /api/auth/verify/request:
    post:
      tags:
      - auth-controller
      operationId: requestVerification
      responses:
        "204":
          description: No Content
      security:
      - bearerAuth: []
  /api/auth/verify/confirm:
    post:
      tags:
      - auth-controller
      operationId: confirmEmail
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRequest"
        required: true
      responses:
        "204":
          description: No Content
  /api/auth/register:
    post:
      tags:
      - auth-controller
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/auth/refresh:
    post:
      tags:
      - auth-controller
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/auth/password/reset:
    post:
      tags:
      - auth-controller
      operationId: resetPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
        required: true
      responses:
        "204":
          description: No Content
  /api/auth/password/forgot:
    post:
      tags:
      - auth-controller
      operationId: forgotPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
        required: true
      responses:
        "204":
          description: No Content
  /api/auth/logout:
    post:
      tags:
      - auth-controller
      operationId: logout
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
        required: true
      responses:
        "204":
          description: No Content
  /api/auth/login:
    post:
      tags:
      - auth-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/auth/google:
    post:
      tags:
      - auth-controller
      operationId: google
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleAuthRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /api/users/me:
    get:
      tags:
      - user-controller
      operationId: getSelf
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserSelfResponse"
      security:
      - bearerAuth: []
    delete:
      tags:
      - user-controller
      operationId: deleteSelf
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteAccountRequest"
      responses:
        "204":
          description: No Content
      security:
      - bearerAuth: []
    patch:
      tags:
      - user-controller
      operationId: updateProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserSelfResponse"
      security:
      - bearerAuth: []
...
  /api/users/{username}:
    get:
      tags:
      - user-controller
      operationId: getProfile
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserProfileResponse"
  /api/users/me/stats:
    get:
      tags:
      - user-controller
      operationId: getStats
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserStatsResponse"
      security:
      - bearerAuth: []
  /api/users/me/records:
    get:
      tags:
      - user-controller
      operationId: getRecords
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExerciseRecordsResponse"
      security:
      - bearerAuth: []
  /api/exercises/{id}/records:
    get:
      tags:
      - exercise-controller
      operationId: records
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PersonalRecordResponse"
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    UserSelfResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        emailVerifiedAt:
          type: string
          format: date-time
        displayName:
          type: string
        bio:
          type: string
        profilePictureUrl:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
          - OTHER
          - PREFER_NOT_TO_SAY
        heightCm:
          type: integer
          format: int32
        weightKg:
          type: number
          format: double
        joinedAt:
          type: string
          format: date-time
...
    TokenRequest:
      type: object
      properties:
        token:
          type: string
          minLength: 1
      required:
      - token
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          maxLength: 30
          minLength: 3
        email:
          type: string
          minLength: 1
        password:
          type: string
          maxLength: 72
          minLength: 8
      required:
      - email
      - password
      - username
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
        expiresIn:
          type: integer
          format: int64
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
          minLength: 1
      required:
      - refreshToken
    ResetPasswordRequest:
      type: object
      properties:
        token:
          type: string
          minLength: 1
        newPassword:
          type: string
          maxLength: 72
          minLength: 8
      required:
      - newPassword
      - token
    ForgotPasswordRequest:
      type: object
      properties:
        email:
          type: string
          minLength: 1
      required:
      - email
    LoginRequest:
      type: object
      properties:
        usernameOrEmail:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
      required:
      - password
      - usernameOrEmail
    GoogleAuthRequest:
      type: object
      properties:
        idToken:
          type: string
          minLength: 1
      required:
      - idToken
...
    UpdateProfileRequest:
      type: object
      properties:
        displayName:
          type: string
          maxLength: 50
          minLength: 0
        bio:
          type: string
          maxLength: 300
          minLength: 0
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
          - OTHER
          - PREFER_NOT_TO_SAY
        heightCm:
          type: integer
          format: int32
        weightKg:
          type: number
          format: double
...
    UserProfileResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        displayName:
          type: string
        bio:
          type: string
        profilePictureUrl:
          type: string
        joinedAt:
          type: string
          format: date-time
    UserStatsResponse:
      type: object
      properties:
        totalWorkouts:
          type: integer
          format: int64
        totalActivities:
          type: integer
          format: int64
        totalDistanceMeters:
          type: number
          format: double
        totalVolumeKg:
          type: number
          format: double
    ExerciseRecordsResponse:
      type: object
      properties:
        exerciseId:
          type: string
          format: uuid
        exerciseName:
          type: string
        records:
          type: array
          items:
            $ref: "#/components/schemas/PersonalRecordResponse"
    PersonalRecordResponse:
      type: object
      properties:
        type:
          type: string
          enum:
          - ONE_REP_MAX
          - MAX_WEIGHT
          - MAX_VOLUME
          - MAX_REPS_AT_WEIGHT
        value:
          type: number
          format: double
        achievedAt:
          type: string
          format: date-time
        workoutSetId:
          type: string
          format: uuid
        contextWeightKg:
          type: number
          format: double
...
    DeleteAccountRequest:
      type: object
      properties:
        password:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      in: header
      scheme: bearer
      bearerFormat: JWT